% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_BarplotUpstreamClasses.R
\name{BarplotUpstreamClasses}
\alias{BarplotUpstreamClasses}
\title{Bar plots of upstream-averaged classes of HYPE sub-basins}
\usage{
BarplotUpstreamClasses(x, type = "landuse")
}
\arguments{
\item{x}{Data frame, containing column-wise class fractions with SUBIDs in first column. Typically a result from 
\code{\link{UpstreamGroupSLCClasses}}}

\item{type}{Character keyword.}

\item{desc}{List, with class description labels imported from a 'description.txt' file . See \code{\link{ReadDescription}} for 
formatting details.}
}
\description{
Function to plot upstream-averaged landscape property classes of one or several sub-basins, e.g. land use or soils, 
as bar plots.
}
\details{
\code{BarplotUpstreamClasses} is a wrapper for \code{\link{barplot}}, with labeling
}
\seealso{
\code{\link{UpstreamGroupSLCClasses}}
\code{\link{barplot}}
}

