% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_CCHypeSingleVar.R
\name{CCHypeSingleVar}
\alias{CCHypeSingleVar}
\title{Pearson correlation coefficient for HypeSingleVar arrays}
\usage{
CCHypeSingleVar(sim, obs, progbar = TRUE, ...)
}
\arguments{
\item{sim}{\code{\link{HypeSingleVar}} array with simulated variable (one or several iterations).}

\item{obs}{\code{\link{HypeSingleVar}} array with observed variable, (one iteration). If several iterations are present 
in the array, only the first will be used.}

\item{progbar}{Logical, if \code{TRUE} progress bars will be printed for main computational steps.}
}
\value{
A two-dimensional array with Pearson CCs for all time series pairs provided in \code{sim} and \code{obs}, with values in the same order 
as the second and third dimension as the input, i.e. \code{[subid, iteration]}.
}
\description{
Pearson correlation coefficient calculation for imported HYPE outputs with single variables for several catchments, i.e. time and 
map files, optionally multiple model runs combined.
}

