% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_CreateOptpar.R
\name{CreateOptpar}
\alias{CreateOptpar}
\title{Create an optpar list}
\usage{
CreateOptpar(x, pars, tasks, comment = "")
}
\arguments{
\item{x}{a list with named vector elements, as an object returned from \code{\link{ReadPar}}.}
}
\description{
\code{CreateOptpar} creates a list representing a HYPE optpar.txt file from an imported par.txt file 
and a selection of parameters.
}

