% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{HypeDataExport}
\alias{HypeDataExport}
\alias{WriteAquiferData}
\alias{WriteBranchData}
\alias{WriteCropData}
\alias{WriteDamData}
\alias{WriteLakeData}
\alias{WritePointSourceData}
\title{Read HYPE data files}
\usage{
WriteAquiferData(x, filename = "AquiferData.txt", digits = 10,
  nsmall = 0L, verbose = T)

WriteBranchData(x, filename = "BranchData.txt")

WriteCropData(x, filename = "CropData.txt", digits = 10, nsmall = 0L,
  verbose = T)

WriteDamData(x, filename = "DamData.txt", digits = 10, nsmall = 0L,
  verbose = T)

WriteLakeData(x, filename = "LakeData.txt", digits = 10, nsmall = 0L,
  verbose = T)

WritePointSourceData(x, filename = "PointSourceData.txt", digits = 10,
  nsmall = 0L, verbose = T)
}
\arguments{
\item{x}{The object to be written, a dataframe as returned from the \code{\link{HypeDataImport}} functions.}

\item{filename}{A character string naming a path and file name to write to. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{digits}{Integer, number of significant digits to export. See \code{\link{format}}.}

\item{nsmall}{Integer, number of significant decimals to export. See \code{\link{format}}.}

\item{verbose}{Logical, display informative warning messages if columns contain \code{NA} values or if character strings are
too long. See Details.}
}
\description{
These are simple convenience wrapper functions to export various HYPE data files from R.
}
\details{
Hype data file imports, simple \code{\link{write.table}} wrappers with formatting options adjusted to match HYPE file 
specifications:

\itemize{
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:lakedata.txt}{LakeData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:damdata.txt}{DamData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:mgmtdata.txt}{MgmtData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt}{AquiferData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:pointsourcedata.txt}{PointSourceData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:glacierdata.txt}{GlacierData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:cropdata.txt}{CropData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:branchdata.txt}{BranchData.txt}
}

In most files, HYPE requires \code{NA}-free input in required columns, but empty values are 
allowed in additional comment columns which are not read by HYPE. Informative warnings will be thrown if \code{NA}s are 
found during export. Character string lengths in comment columns of HYPE data files are restricted to 100 characters, 
the functions will return with a warning if longer strings were exported.
}
\examples{
\dontrun{WriteLakeData(mylakedata, "LakeData.txt")}
\dontrun{WriteDamData(mydamdata, "DamData.txt")}
\dontrun{WriteMgmtData(mymgmtdata, "MgmtData.txt")}

}

