% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{HypeImport}
\alias{HypeImport}
\alias{ReadDamData}
\alias{ReadLakeData}
\title{Read HYPE data files}
\usage{
ReadLakeData(filename = "LakeData.txt")

ReadDamData(filename = "DamData.txt")
}
\arguments{
\item{filename}{Path to and file name of the LakeData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}
}
\value{
\code{ReadLakeData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a LakeData file as data frame into R.
}
\details{
\code{ReadLakeData} is a simple \code{\link{read.table}} wrapper, mainly added to provide a comparable 
function to the other RHYPE imports.
}
\examples{
\dontrun{ReadLakeData("LakeData.txt")}

}

