% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HypeMultiVar_unfinished!.R
\name{HypeMultiVar}
\alias{HypeMultiVar}
\title{HypeMultiVar arrays}
\usage{
HypeMultiVar(x, variable = NULL, subid, hype.var)
}
\arguments{
\item{x}{numeric \code{\link{data.frame}} with \code{\link{POSIXct}} formatted time steps in the first, and \code{\link{numeric}} 
variables in the remaining columns.}

\item{variable}{Character vector of four-letter keywords to specify HYPE variable IDs, corresponding to second to 
last column in \code{x}. Defaults to column names for second to last column.}

\item{subid}{Integer with HYPE sub-basin ID, corresponding to second to last column in \code{x}.}

\item{hype.var}{Character string, four-letter keyword to specify HYPE variable ID, see 
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:info.txt:variables}{list of HYPE variable}. 
Not case-sensitive.}
}
\value{
Returns a 3-dimensional array with 
\code{[time, subid, iteration]} dimensions and additional \code{\link{attributes}}:
\describe{
\item{\strong{date}}{A vector of date-times. Corresponds to 1st array dimension.}
\item{\strong{subid}}{A vector of SUBIDs. Corresponds to 2nd array dimension.}
\item{\strong{variable}}{HYPE output variable ID.}
}
}
\description{
Constructor function for data frames which hold equidistant time series of multiple HYPE variables for a single sub-basin, 
typically imported HYPE basin output results.
}
\details{
S3 class constructor function for array objects which can hold (multiple) HYPE time or map output results.
}
\examples{
\dontrun{HypeSingleVar(mytimeoutput, date = mydates, subid = c(23, 45, 56), hype.var = "cctn")}

}

