% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_HypeSubidChecks.R
\name{HypeSubidChecks}
\alias{HypeSubidChecks}
\alias{IsRegulated}
\alias{IsHeadwater}
\title{Check HYPE SUBID properties}
\usage{
IsRegulated(subid, dd = NULL, ld = NULL)

IsHeadwater(subid, gd)
}
\description{
Quickly query single SUBIDs for various properties.
}
