% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYPEtools-package.R
\docType{package}
\name{HYPEtools-package}
\alias{HYPEtools}
\alias{HYPEtools-package}
\title{HYPEtools: Tools for Processing and Analyzing Files from the Hydrological Catchment Model HYPE}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Work with model files (setup, input, output) from the hydrological catchment model HYPE: Streamlined file import and export, standard evaluation plot routines, diverse post-processing and aggregation routines for hydrological model analysis. The HYPEtools package is also archived at \doi{10.5281/zenodo.7627955} and can be cited in publications with Brendel et al. (2024) \doi{10.1016/j.envsoft.2024.106094}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hypeweb.smhi.se/}
  \item \url{https://github.com/rcapell/HYPEtools}
  \item Report bugs at \url{https://github.com/rcapell/HYPEtools/issues}
}

}
\author{
\strong{Maintainer}: Rene Capell \email{hypetools.rene@smhi.se} (\href{https://orcid.org/0000-0002-7784-1313}{ORCID})

Authors:
\itemize{
  \item Rene Capell \email{hypetools.rene@smhi.se} (\href{https://orcid.org/0000-0002-7784-1313}{ORCID})
  \item Conrad Brendel \email{conrad.brendel@smhi.se} (\href{https://orcid.org/0000-0002-5199-0580}{ORCID})
}

Other contributors:
\itemize{
  \item Jafet Andersson [contributor]
  \item David Gustafsson [contributor]
  \item Jude Musuuza [contributor]
  \item Jude Lubega [contributor]
}

}
\keyword{internal}
