% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHYPE-package.R
\docType{package}
\name{HYPEtools}
\alias{HYPEtools}
\alias{HYPEtools-package}
\title{HYPEtools}
\description{
This package contains functions to handle and analyse input and
output files from the conceptual rainfall-runoff model HYPE (HYdrological
Predictions for the Environment).
}
