% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_LakeDataOutlets.R
\name{LakeDataOutlets}
\alias{LakeDataOutlets}
\title{Identify LakeData Outlet SUBIDs}
\usage{
LakeDataOutlets(ld, gd)
}
\arguments{
\item{ld}{A data frame containing LakeData information (i.e. an imported 'LakeData.txt' file).
Must contain the columns \code{LAKEID} and \code{SUBID}.}

\item{gd}{A data frame containing GeoData information used to determine
the upstream connectivity between SUBIDs (i.e. an imported 'GeoData.txt' file).}
}
\value{
An integer vector containing the SUBIDs identified as lake
outlets. One or more SUBIDs may be returned for a lake if the lake
contains multiple downstream outlets.
}
\description{
Identifies the SUBID(s) corresponding to the outlet of each lake in a
LakeData data frame.
}
\details{
A SUBID is identified as a lake outlet when all SUBIDs belonging to
the same lake occur in the upstream area of that SUBID. The function
uses \code{\link[=AllUpstreamSubids]{AllUpstreamSubids()}} to determine the upstream SUBIDs.
}
\examples{
ld <- ReadLakeData(
  filename = system.file("demo_model", "LakeData.txt", package = "HYPEtools")
)
gd <- ReadGeoData(
  filename = system.file("demo_model", "GeoData.txt", package = "HYPEtools")
)

LakeDataOutlets(ld, gd)

}
\seealso{
\code{\link[=AllUpstreamSubids]{AllUpstreamSubids()}}
}
