% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the shinyAppDemo app}
\usage{
launchApp(option.x = NULL, option.map = NULL, option.var.name = NULL)
}
\value{
shiny application object
}
\description{
launches the shinyAppDemo app
}
