% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HypeGeoData.R
\name{merge}
\alias{merge}
\alias{merge.HypeGeoData}
\title{Merge HypeGeoData object}
\usage{
merge.HypeGeoData(x, y, sort = FALSE, ...)
}
\arguments{
\item{x}{\code{link{HypeGeoData}} data frame, HYPE GeoData table to be extended with new columns.}

\item{y}{Data frame, with mandatory \code{SUBID} column.}

\item{sort}{Logical, result sorting by \code{by} columns. Defaults to \code{FALSE}, as opposed to default method.}

\item{...}{Arguments passed to S3 method for data frames, see \code{\link{merge}} and Details.}
}
\value{
A \code{HypeGeoData} data frame.
}
\description{
Merge an imported HYPE GeoData table of class \code{link{HypeGeoData}} with another data frame.
}
\details{
\code{merge.HypeGeoData} allows to merge new columns to an existing HYPE GeoData table, while preserving the \code{HypeGeoData} 
class attribute. Duplicate columns are marked with a \code{".y"}-suffix for the merged \code{y} data frame.

The following arguments of the default method are hard-coded:

\itemize{
\item{\code{by, by.x, by.y}, set to \code{"SUBID"}}
\item{\code{suffixes}, set to \code{c("", ".y")}}
}

The method warns if any of these arguments is supplied by the user. To override these nonetheless, call the data frame method 
explicitly (\code{merge.data.frame}).
}
\seealso{
\code{\link[base]{merge}}, the S3 generic function.
}
