% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MergeXobs.R
\name{MergeXobs}
\alias{MergeXobs}
\title{Merge two Xobs data frames}
\usage{
MergeXobs(x, y, x.first = TRUE, comment = "")
}
\arguments{
\item{x, y}{Data frames with additional attributes \code{comment}, \code{variable}, and \code{subid}, 
typically imported using \code{\link{ReadXobs}}. For details on attribute format, see there.}

\item{x.first}{Logical, indicating which input data frame will be given precedence in case of common 
columns (identical observation variable and SUBID combinations in \code{x} and \code{y}). If \code{TRUE}, 
values from columns in \code{x} will take precedence, and values from \code{y} will only be added if 
\code{x} values are missing. Vice versa if \code{FALSE}.}

\item{comment}{Character string, will be added to the result as attribute \code{comment}. If empty, 
comment attributes from \code{x} and \code{y} will be merged to new comment string.}
}
\value{
\code{MergeXobs} returns a data frame with attributes for Xobs data.
}
\description{
Function to merge two Xobs data frames, with handling of overlapping time periods and time periods gaps 
as well as merging of common columns.
}
\details{
\code{MergeXobs} handles time steps of different lengths (e.g. daily, hourly), but requires identical time 
step lengths from both input data frames.
}
\examples{
\dontrun{MergeXobs(x = myxobs1, y = myxobs2)}
}

