% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hydroGOF.R
\name{NSE.HypeSingleVar}
\alias{NSE}
\alias{NSE.HypeSingleVar}
\title{Nash-Sutcliffe Efficiency}
\usage{
\method{NSE}{HypeSingleVar}(sim, obs, na.rm = TRUE, progbar = TRUE, ...)
}
\arguments{
\item{sim}{\code{\link{HypeSingleVar}} array with simulated variable.}

\item{obs}{\code{\link{HypeSingleVar}} array with observed variable.}

\item{na.rm}{
a logical value indicating whether 'NA' should be stripped before the computation proceeds. \cr
When an 'NA' value is found at the i-th position in \code{obs} \bold{OR} \code{sim}, the i-th value of \code{obs} \bold{AND} \code{sim} are removed before the computation.
}

\item{progbar}{Logical, if \code{TRUE} progress bars will be printed for main computational steps.}

\item{...}{ignored}
}
\value{
A two-dimensional array with NSEs for all time series pairs provided in \code{sim} and \code{obs}, with values in the same order 
as the second and third dimension as the input, i.e. \code{[subid, iteration]}.
}
\description{
Nash-sutcliffe Efficiency calculation for imported HYPE outputs with single variables for several catchments, i.e. time and 
map files, optionally multiple model runs combined.
}

