% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hydroGOF.R
\name{NSE}
\alias{NSE}
\alias{NSE.HypeSingleVar}
\title{Nash-Sutcliffe Efficiency}
\usage{
\method{NSE}{HypeSingleVar}(sim, obs, na.rm = TRUE, progbar = TRUE, ...)
}
\arguments{
\item{sim}{\code{link{HypeSingleVar}} array with simulated variable.}

\item{obs}{\code{link{HypeSingleVar}} array with observed variable.}

\item{progbar}{Logical, if \code{TRUE} progress bars will be printed for main computational steps.}

\item{...}{ignored}
}

