% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_OptimisedClasses.R
\name{OptimisedClasses}
\alias{OptimisedClasses}
\title{Get optimised classes from an imported optpar.txt file}
\usage{
OptimisedClasses(x)
}
\arguments{
\item{x}{list with named elements, as an object returned from \code{\link{ReadOptpar}}.}
}
\value{
\code{OptimisedClasses} returns a named list with one vector element for each parameter found in \code{x}. List element 
names are HYPE parameter names. Each vector contains the optimised class numbers for the respective parameter.

\item \code{calib}, a list of vectors with (soil or land use) class numbers of parameters included in calibration (parameters 
with identical min and max values are omitted in calibration, but need to be specified in optpar files). Parameter names 
as list element names.
}
\description{
\code{OptimisedClasses} checks which classes (land use or soil) of parameters in an imported optpar list are actually 
optimised, i.e. have a min/max range larger than zero.
}
\details{
\code{OptimisedClasses} allows to quickly check which classes of parameters in an optpar.txt file are actually optimised 
during a HYPE otimisation run. The function compares min and max values in the \code{pars} element of an imported HYPE 
optpar.txt file to identify those.
}

