% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PartyParrot.R
\name{PartyParrot}
\alias{PartyParrot}
\title{Create a Party Parrot.}
\usage{
PartyParrot(sound = 8)
}
\arguments{
\item{sound}{Character string or number specifying which sound to play when showing the Party Parrot. See the \code{beep} function in the \code{beepr} package.}
}
\value{
Returns a Party Parrot to the Console.
}
\description{
Creates a Party Parrot.
}
\details{
\code{PartyParrot} generates a Party Parrot. Uses for Party Parrots include, for example, celebrating the successful execution of a script.
}
\examples{
PartyParrot()

}
