% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hydroGOF.R
\name{pbias}
\alias{pbias}
\alias{pbias.HypeSingleVar}
\title{Percent bias}
\usage{
\method{pbias}{HypeSingleVar}(sim, obs, na.rm = TRUE, progbar = TRUE, ...)
}
\arguments{
\item{sim}{\code{\link{HypeSingleVar}} array with simulated variable (one or several iterations).}

\item{obs}{\code{\link{HypeSingleVar}} array with observed variable, (one iteration). If several iterations are present
in the array, only the first will be used.}

\item{na.rm}{Logical. If \code{TRUE}, incomplete sim-obs pairs will be removed prior to PBIAS computation.}

\item{progbar}{Logical. If \code{TRUE}, progress bars will be printed for main computational steps.}

\item{...}{ignored}
}
\value{
\code{pbias.HypeSingleVar} returns a 2-dimensional array of NSE performances for all SUBIDs and model iterations provided in 
argument \code{sim}, with values in the same order 
as the second and third dimension in \code{sim}, i.e. \code{[subid, iteration]}.
}
\description{
Percent bias (PBIAS) calculation for imported HYPE outputs with single variables for several catchments, i.e. time and 
map files, optionally multiple model runs combined.
}
\seealso{
\code{\link[hydroGOF]{pbias}}, the S3 generic function defined in package 'hydroGOF'.
}

