% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PlotBasinSummary.R
\name{PlotBasinSummary}
\alias{PlotBasinSummary}
\title{Plot a summary of model results for a single sub-basin}
\usage{
PlotBasinSummary(x, filename = NULL, gd = gd, bd = NULL, gcl = gcl,
  psd = NULL, desc = NULL, from = 1, to = nrow(x))
}
\arguments{
\item{x}{Data frame, with column-wise daily time series of HYPE variables. Date-times in 
\code{\link{POSIXct}} format in first column. Typically an imported basin output file from HYPE using \code{\link{ReadBasinOutput}}. 
See details for HYPE output variables required for plotting.}

\item{filename}{String, file name for plotting to \code{\link{png}} device. \code{NULL}, the default, triggers a plot on a new
screen device. \emph{Device dimensions are currently hard-coded.}}

\item{timestep}{Character string, timestep of \code{x}, one of \code{"month"}, \code{"week"}, \code{"day"}, or 
\code{"nhour"} (n = number of hours). If not provided, an attribute \code{timestep} is required in \code{x}.}
}
\description{
Plot a standard suite of plots summarising properties of a sub-basin including upstream area and model performance 
for discharge and nutrients.
}
\details{

}

