% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PlotEvaluationMaps.R
\name{PlotEvaluationMaps}
\alias{PlotEvaluationMaps}
\title{Function for plotting up to two HYPE simulation results in maps.}
\usage{
PlotEvaluationMaps(
  tempOutDirectory,
  figuresDirectory = NULL,
  referenceSubassFile,
  simulationSubassFile,
  subbasinShapeFile,
  geodataFile,
  simulationInfoFile,
  streamShapeFile = NULL,
  criterion = c("NSE", "KGE", "CC", "MAE", "RE(\%)", "RMSE", "KGESD", "KGEM"),
  visualization = c("best.simulation", "relative.difference", "comparison"),
  data.presentation = c("outlets", "polygons", "centroids"),
  evaluation.dataset,
  simulation.names = NULL,
  marker.size = NULL,
  show.borders = FALSE,
  show.streams = FALSE,
  histogram.fill = NULL,
  used.colours = NULL,
  domain.name,
  gauge.list = NULL
)
}
\arguments{
\item{tempOutDirectory}{Temporary directory for intermediate data e.g., geo-spatial objects. It is mandatory character string that is also used for figures if the figures directory is not specified. To save time, intermediate data is saved the first time functions are called and loaded from disk during subsequent calls (e.g., to visualize different performance metrics with the same variable).}

\item{figuresDirectory}{The figures output directory, an optional character string. If not set, figures are saved to the temporary directory with the intermediate objects.}

\item{referenceSubassFile}{A mandatory character string specifying the reference (default) HYPE simulation subass file.}

\item{simulationSubassFile}{An optional character string specifying the second simulation subass file e.g., for comparison to the reference or for computing relative differences between simulation performances.}

\item{subbasinShapeFile}{The path to the shape file containing the sub-basin geometries. A mandatory character string.}

\item{geodataFile}{A mandatory character string specifying the path to the GeoData.txt file, which provides spatial information of the sub-basins. The outlets of the gauged sub-basins in the subass files are also read from this file and converted into geo-spatial objects for visualization.}

\item{simulationInfoFile}{A mandatory character string specifying the path to the info.txt file containing the simulation settings, e.g., start/end dates, aggregation periods, etc.}

\item{streamShapeFile}{An optional character string for the path to the shape file containing the stream network of the model domain, which  should be projected to the WGS84 system for consistency. If the stream network is desired without specifying the shape file, the 50m-resolution map from Natural Earth  is used (requires network connection). The layer may not provide the desired level of detail however.}

\item{criterion}{An optional valid name of a HYPE subass evaluation criterion. 'NSE', 'KGE', 'CC', 'MAE', 'RE(\%)', and 'RMSE' are currently implemented. NSE is plotted by default.}

\item{visualization}{The mandatory type of visualization from "relative.difference", "best.simulation" and "comparison". The first two options produce a single map while the third produces two side-by-side maps. "Comparison" and "relative difference" require two simulations.}

\item{data.presentation}{The mandatory data presentation mode on maps, one of "polygons", "outlets" or "centroids". "Polygons" are useful for spatially distributed data e.g., evaporation. "Outlets" and "centroids" are useful for point observations, e.g., discharge and sediment plotted at the sub-basin outlets and centroids, respectively.}

\item{evaluation.dataset}{A mandatory character string for the name of data being analysed, e.g., "Discharge", "Actual ET", "Snow Water Equivalent", etc.}

\item{simulation.names}{A vector of two (at least one) character strings with the "names" of the simulations e.g., model versions, used as titles on the maps. The first item mandatory.}

\item{marker.size}{A numerical value for the marker size, optional.}

\item{show.borders}{Logical choice to show political borders (default is FALSE). If requested, the 110m-resolution map will be downloaded from the Natural Earth portal, which requires an internet connection.}

\item{show.streams}{Logical choice to show river network on the map (default is FALSE).}

\item{histogram.fill}{A character string of length two for the histogram fill colours. In case of a single map, the first colour (mandatory) is used.}

\item{used.colours}{An optional vector of colours for the maps.}

\item{domain.name}{The name of the domain. This character string is mandatory. The global model is called "wwhype".}

\item{gauge.list}{A subset of gauges to be plotted (must exist in the subass file), if some gauges in the subass file are to be excluded.}
}
\description{
Draws maps for selected HYPE variables and performance metrics with pretty scale discretizations and colours.
}
