% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_r.R
\name{r}
\alias{r}
\alias{r.HypeSingleVar}
\title{Pearson product-moment correlation coefficient r}
\usage{
r(sim, obs, ...)

\method{r}{HypeSingleVar}(sim, obs, progbar = TRUE, ...)
}
\arguments{
\item{sim}{\code{\link[=HypeSingleVar]{HypeSingleVar array}} with simulated variable (one or several iterations).}

\item{obs}{\code{\link[=HypeSingleVar]{HypeSingleVar array}} with observed variable, (one iteration). If several iterations 
are present in the array, only the first will be used.}

\item{...}{Ignored.}

\item{progbar}{Logical, if \code{TRUE} progress bars will be printed for main computational steps.}
}
\value{
\code{r.HypeSingleVar} returns a 2-dimensional array of Pearson correlation coefficients for all SUBIDs and model 
iterations provided in argument \code{sim}, with values in the same order 
as the second and third dimension in \code{sim}, i.e. \code{[subid, iteration]}.
}
\description{
Pearson product-moment correlation coefficient calculation, a specific case of function \code{\link{cor}}.
}
\details{
This function wraps a call to \code{cor(x = obs, y = sim, use = "na.or.complete", method = "pearson")}. 

Method \code{r.HypeSingleVar} calculates Pearson's r for imported HYPE outputs with single variables for several 
catchments, i.e. time and map files, optionally multiple model runs combined, typically results from calibration runs.
}
\seealso{
\code{\link{cor}}, on which the function is based. \code{\link{ReadWsOutput}} for importing HYPE calibration results.
}
