% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadAllsim}
\alias{ReadAllsim}
\title{Read an 'allsim.txt' file}
\usage{
ReadAllsim(filename = "allsim.txt")
}
\arguments{
\item{filename}{Path to and file name of the 'allsim.txt' file to import.}
}
\value{
\code{ReadAllsim} returns a data frame.
}
\description{
This is a convenience wrapper function to import an allsim.txt optimisation result file as data frame into R.
}
\details{
\code{ReadAllsim} is just \code{read.table(file = filename, header = T, sep = ",")}, mainly added to provide a 
comparable function to the other RHYPE imports.
}
\examples{
\dontrun{ReadAllsim("allsim.txt")}

}

