% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadAquiferData}
\alias{ReadAquiferData}
\title{Read an 'AquiferData.txt' file}
\usage{
ReadAquiferData(filename = "AquiferData.txt", sep = "\\t")
}
\arguments{
\item{filename}{Path to and file name of the AquiferData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{sep}{Character string, field separator as in \code{\link{read.table}}.}
}
\value{
\code{ReadAquiferData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a HYPE AquiferData file as data frame into R.
}
\details{
\code{ReadAquiferData} is a simple \code{\link{read.table}} wrapper, mainly added to provide a comparable 
function to other RHYPE import functions. Will check for \code{NA} values in imported data and return a warning if any are found. 
HYPE requires \code{NA}-free input in required 'AquiferData.txt' columns, but empty values are allowed in comment columns 
which are not read.
}
\examples{
\dontrun{ReadAquiferData("../myhype/AquiferData.txt")}

}

