% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadCropData}
\alias{ReadCropData}
\title{Read a 'CropData.txt' file}
\usage{
ReadCropData(filename = "CropData.txt")
}
\arguments{
\item{filename}{Path to and file name of the CropData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}
}
\value{
\code{ReadCropData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a CropData file as data frame into R.
}
\details{
\code{ReadCropData} is just \code{read.table(file = filename, header = T)}, mainly added to provide a comparable 
function to the other RHYPE imports.
}
\examples{
\dontrun{ReadCropData("CropData.txt")}

}

