% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadGlacierData}
\alias{ReadGlacierData}
\title{Read a 'GlacierData.txt' file}
\usage{
ReadGlacierData(filename = "GlacierData.txt", sep = "\\t")
}
\arguments{
\item{filename}{Path to and file name of the GlacierData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{sep}{character string. Field separator character as described in \code{\link{read.table}}.}
}
\value{
\code{ReadGlacierData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a GlacierData file as data frame into R.
}
\details{
\code{ReadGlacierData} runs \code{read.table(file = filename, header = T, sep = sep)} and forces upper-case column names.
}
\examples{
\dontrun{ReadGlacierData("GlacierData.txt")}

}

