% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadOptpar}
\alias{ReadOptpar}
\title{Read an 'optpar.txt' file}
\usage{
ReadOptpar(filename)
}
\arguments{
\item{filename}{Path to and file name of the 'optpar.txt' file to import.}
}
\value{
\code{ReadOptpar} returns a \code{\link{list}} object with four elements: \itemize{ 
\item \code{comment}, the file's first-row comment string.
\item \code{tasks}, a two-column dataframe with row-wise key-value pairs for tasks and settings.
\item \code{pars}, a list of dataframes, each containing values for one parameter. Three columns each, holding parameter 
range minima, maxima, and intervals. 
The number of rows in each dataframe corresponds to the number of soil or land use classes for class-specific parameters. 
Parameter names as list element names.
\item \code{calib}, a list of vectors with (soil or land use) class numbers of parameters included in calibration (parameters 
with identical min and max values are omitted in calibration, but need to be specified in optpar files). Parameter names 
as list element names.
}
}
\description{
This function imports an 'optpar.txt' into a list.
}
\details{
\code{ReadOptpar} imports a HYPE 'optpar.txt' file. Optpar files contain instructions for parameter calibration/optimisation 
and parameter value ranges, for details on the file format, see the
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:optpar.txt}{optpar.txt online documentation}.
}
\seealso{
\code{\link{ReadPar}}
}

