% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadPmsf}
\alias{ReadPmsf}
\title{Read a 'pmsf.txt' file}
\usage{
ReadPmsf(filename = "pmsf.txt")
}
\arguments{
\item{filename}{Path to and file name of the pmsf file to import. Windows users: Note that
Paths are separated by '/', not '\\'.}
}
\value{
\code{ReadPmsf} returns an integer vector.
}
\description{
This is a small convenience function to import a 'partial model setup file' as integer vector into R.
}
\details{
Pmsf.txt files imported with \code{ReadPmsf} are stripped from the first value containing the total number of subcatchments
in the file. No additional attribute is added to hold this number since it can be easily obtained using \code{\link{length}}.
}
\examples{
\dontrun{ReadLakeData("pmsf.txt")}
}

