% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadPointSourceData}
\alias{ReadPointSourceData}
\title{Read a 'PointSourceData.txt' file}
\usage{
ReadPointSourceData(filename = "PointSourceData.txt")
}
\arguments{
\item{filename}{Path to and file name of the PointSourceData file to import.}
}
\value{
\code{ReadPointSourceData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a PointSourceData file as data frame into R.
}
\details{
\code{ReadPointSourceData} is just \code{read.table(file = filename, header = T)}, mainly added to provide a comparable
function to the other RHYPE imports.
}
\examples{
\dontrun{ReadPointSourceData("PointSourceData.txt")}
}

