% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadSimass}
\alias{ReadSimass}
\title{Read a 'simass.txt' file}
\usage{
ReadSimass(filename = "simass.txt")
}
\arguments{
\item{filename}{Path to and file name of the 'simass.txt' file to import.}
}
\value{
\code{ReadSubass} returns a data frame with column-wise performance measure values of evaluated variable pairs and performance 
measure names in the first column. Additional attributes: 
\itemize{
\item{\code{n.data.regional}, \code{numeric} vector with number of data points used in regional criteria (optional, depending on 
HYPE version)}
\item{\code{n.area.mean}, \code{numeric} vector with number of subbasins used in mean and median criteria (optional, depending on 
HYPE version)}
\item{\code{variable.obs}, \code{character} vector with HYPE variable ID(s) used as observed variable}
\item{\code{variable.sim}, \code{character} vector with HYPE variable ID(s) used as simulated variable}
\item{\code{agg.period}, \code{numeric} vector with integer code(s) for time period aggregation used in evaluation. Same format 
as used in info.txt files, i.e. 1 = daily, 2 = weekly, 3 = monthly, 4 = annual.}
\item{\code{n.simulation}, \code{integer}, simulation number (e.g. with Monte Carlo simulations)}
\item{\code{crit.total}, \code{numeric}, total criteria value}
\item{\code{crit.conditional}, \code{numeric}, conditional criteria value}
\item{\code{threshold}, \code{integer}, data limit threshold}
}
}
\description{
This is a convenience wrapper function to import an simass.txt simulation assessment file as data frame into R.
Simulation assessment files contain domain-wide aggregated performance criteria results, as defined in 'info.txt'.
}
\details{
\code{ReadSimass} imports a simulation assessement file into R. Evaluated variables and other metadata are stored as additional 
\code{\link{attributes}}.

The function interpretes character-coded time steps (e.g. \code{"DD"} for daily time steps), as used in some HYPE versions. 
Sub-daily time steps are currently not treated and will probably result in a warning during time step evaluation within the 
function. Please update issue #106 on \href{https://github.com/rcapell/HYPEtools} if you need support for sub-daily time steps!
}
\examples{
\dontrun{ReadSimass("simass.txt")}

}
