% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ReadWsOutput.R
\name{ReadWsOutput}
\alias{ReadWsOutput}
\title{Read optimisation simulation results}
\usage{
ReadWsOutput(path, type = c("time", "map", "basin"), hype.var = NULL,
  subid = NULL, dt.format = NULL, subid.select = NULL, from = NULL,
  to = NULL, progbar = T)
}
\arguments{
\item{path}{Character string, path to the directory holding simulation output files to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{type}{Character string, keyword for HYPE output file type to import. One of \code{"time"}, \code{"map"}, or 
\code{"basin"}. Can be abbreviated. The first two require specification of argument \code{hype.var}, the latter of argument 
\code{subid}. Format of return value depends on output type, see details.}

\item{hype.var}{Character string, keyword to specify HYPE output variable to import. Not case-sensitive. Required in combination 
with \code{type} \code{"time"} or \code{"map"}.}

\item{subid}{Integer, giving a single SUBID for which to import basin output files. Required in combination with \code{type} 
\code{"basin"}.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}, for conversion of date-time information in imported 
result files to POSIX dates, which are returned as attribute. Incomplete format strings for monthly and annual values allowed, e.g. 
'\%Y'. Defaults to \code{NULL}, which prevents date-time conversion, applicable e.g. for files containing just one column of 
summary values over the model period.}

\item{subid.select}{Integer, vector of HYPE SUBIDs to select from imported time and map output files.}

\item{from}{Integer. For partial imports, number of simulation iteration to start from.}

\item{to}{Integer. For partial imports, number of simulation iteration to end with.}

\item{progbar}{Logical, display a progress bar while importing HYPE output files. Adds overhead to calculation time but useful 
when many files are imported.}
}
\value{
\code{ReadWsOutput} returns a 3-dimensional array with additional attributes. The array content depends on the HYPE output file type 
specified in argument \code{type}. Time and map output file imports return an array of class \code{\link{HypeSingleVar}} with 
\code{[time, subid, iteration]} dimensions, basin output file imports return a numeric array with \code{[time, subid, iteration]} 
dimensions.

Returned arrays contain additional \code{\link{attributes}}:
\describe{
\item{\strong{date}}{A vector of date-times, \code{POSIX} if argument \code{dt.format} is non-\code{NULL}. Corresponds to 1st array 
dimension.}
\item{\strong{subid}}{A vector of SUBIDs. Corresponds to 2nd array dimension for time and map output files.}
\item{\strong{variable}}{A vector of HYPE output variables. Corresponds to 2nd array dimension for basin output files.}
}
}
\description{
Read and combine HYPE optimisation simulation output files, generated with 'task WS' during HYPE optimisation runs. Outputs can 
consist of basin, time, or map output files.
}
\details{
HYPE optimisation routines optionally allow for generation of simulation output files for each iteration in the optimisation routine. 
For further details see decumentation on 'task WS' in the 
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:optpar.txt}{optpar.txt online documentation}. 

\code{ReadWsOutput} imports and combines all simulation iterations in an \code{\link{array}}, which can then be easily used in 
further analysis, most likely in combination with performance and parameter values from an imported corresponding 'allsim.txt' file. 
The result folder containing HYPE WS results, argument \code{path}, can contain other files as well, \code{ReadWsOutput} searches for 
file name pattern to filter targeted result files.

For large numbers of result files, simulations can be partially imported using arguments \code{from} and \code{to}, in order to avoid 
memory exceedance problems.
}
\examples{
\dontrun{ReadWsOutput(path = "../my_optim_results/", hype.var = "cout", dt.format = "\%Y-\%m")}
}

