% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHYPE.R
\docType{package}
\name{RHYPE}
\alias{RHYPE}
\alias{RHYPE-package}
\title{RHYPE}
\description{
This package contains functions to handle and analyse input and
output files from the conceptual rainfall-runoff model HYPE (HYdrological
Predictions for the Environment).
}

