% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ScaleAquiferData.R
\name{ScaleAquiferData}
\alias{ScaleAquiferData}
\title{Scale 'AquiferData.txt' files to different model time steps}
\usage{
ScaleAquiferData(
  x = NULL,
  timestep.ratio = 1/24,
  digits = 3,
  verbose = TRUE,
  print.par = FALSE
)
}
\arguments{
\item{x}{Data frame containing HYPE AquiferData contents. Typically imported with \code{\link[=ReadAquiferData]{ReadAquiferData()}}.}

\item{timestep.ratio}{Numeric, time step scaling factor. Defaults to (1/24) to scale from daily to hourly time steps. To scale from hourly to daily time steps use 24.}

\item{digits}{Integer, number of significant digits in scaled parameter values to export. See \code{\link[=signif]{signif()}}.}

\item{verbose}{Logical, if \code{TRUE}, then information will be printed.}

\item{print.par}{Logical, print known time-scale dependent recession coefficients instead of scaling an AquiferData data frame.}
}
\value{
A \code{\link[=dataframe]{dataframe()}} object as supplied in \code{x}, with re-scaled recession coefficients, or nothing if \code{print.par = TRUE}.
}
\description{
\code{ScaleAquiferData} scales the \code{RETRATE} time step-dependent recession coefficient in an imported
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt}{HYPE 'AquiferData.txt'} file to a
new target time step. See HYPE wiki tutorial on \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_tutorials:subdaily_timesteps}{sub-daily time steps}.
}
\details{
\code{ScaleAquiferData} applies a user-specified scaling factor \code{timestep.ratio} to the \code{RETRATE} time step-dependent recession coefficient
in a HYPE AquiferData data frame. All \code{RETRATE} values that are not equal to zero are assumed to be aquifer rows and will be converted to the new time step.
Recession coefficients are matched against an inbuilt set of column names. To see these names, call \code{ScaleAquiferData(print.par = TRUE)}.
\emph{\href{https://github.com/rcapell/HYPEtools/issues}{Please notify us} if you find any missing coefficients.}

Timestep-dependent recession coefficients are scaled using the relationship described in:
Nalbantis, Ioannis (1995). “Use of multiple-time-step information in rainfall-runoff modelling”, Journal of Hydrology 165, 1-4, pp. 135–159.

\code{new_coefficient_value = 1 - (1 - old_coefficient_value)^time_step_ratio}
}
\examples{
# Import daily HYPE AquiferData file
ad <- ReadAquiferData(filename = system.file("demo_model", "AquiferData.txt", package = "HYPEtools"))
# Scale to hourly time steps
ScaleAquiferData(x = ad)
# Print all time scale-dependent coefficients known to the function
ScaleAquiferData(print.par = TRUE)

}
\seealso{
\code{\link{ScalePar}}
}
