% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_UpstreamCropids.R
\name{UpstreamCropids}
\alias{UpstreamCropids}
\title{Calculate area-weighted upstream averages of CropID fractions}
\usage{
UpstreamCropids(subid = NULL, gd, gc, bd = NULL, signif.digits = 3,
  progbar = T)
}
\arguments{
\item{subid}{Integer vector of SUBIDs for which to calculate upstream properties (must exist in \code{gd}).
If \code{NULL} (default), upstream areas for all SUBIDs will be calculated.}

\item{gd}{A data frame containing a column with SUBIDs and a column with areas, e.g. an imported 'GeoData.txt' file.}

\item{gc}{A data frame containing an imported 'GeoClass.txt' file.}

\item{bd}{A data frame with bifurcation connections, e.g. an imported 'BranchData.txt' file. Optional argument.}

\item{signif.digits}{Integer, number of significant digits to round upstream SLCs to. See also \code{\link{signif}}. Set to \code{NULL} to prevent rounding.}

\item{progbar}{Logical, display a progress bar while calculating SLC class fractions. Adds overhead to calculation time but useful when \code{subid}
is \code{NULL} or contains many SUBIDs.}
}
\value{
\code{UpstreamCropids} returns a data frame with SUBIDs in the first column, and upstream CropID fractions in the following columns.
}
\description{
Function to calculate averages of CropID fractions calculated from imported GeoData.txt and GeoClass.txt files.
}
\details{
\code{UpstreamCropids} calculates area-weighted upstream averages of CropID fractions from SLC class fractions in a GeoData table and corresponding
CropIDs in a GeoClass table, including branch connections in case of stream bifurcations but not including potential irrigation links or
groundwater flows. Averages are weighted by sub-catchment area.
}
\note{
\code{UpstreamCropids} expects SLC class columns in argument \code{gd} to be ordered in ascending order.
}
\examples{
\dontrun{UpstreamCropids(subid = 21, gd = mygeodata, gc = mygeoclass, bd = mybranchdata)}
}
\seealso{
\code{\link{UpstreamSLCClasses}}
\code{\link{UpstreamGeoData}}
\code{\link{SumUpstreamArea}}
\code{\link{AllUpstreamSubids}}
}

