% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_UpstreamPointSources.R
\name{UpstreamPointSources}
\alias{UpstreamPointSources}
\title{Summarize point source emissions of all upstream areas}
\usage{
UpstreamPointSources(subid = NULL, gd, psd, bd = NULL, signif.digits = 4,
  progbar = T)
}
\arguments{
\item{subid}{Integer vector of SUBIDs to calculate upstream SUBID fractions for (must exist in \code{gd}).
If \code{NULL}, upstream areas for all SUBIDs will be calculated.}

\item{gd}{A data frame containing columns 'SUBID' with SUBIDs and 'MAINDOWN' with downstream SUBIDs,
e.g. an imported 'GeoData.txt' file.}

\item{psd}{A data frame with HYPE point source specifications, e.g. and imported 'PointSourceData.txt' file.}

\item{bd}{A data frame with bifurcation connections, e.g. an imported 'BranchData.txt' file. Optional argument.}

\item{signif.digits}{Integer, number of significant digits to round upstream SLCs to. See also \code{\link{signif}}. Set to \code{NULL} to prevent rounding.}

\item{progbar}{Logical, display a progress bar while calculating SLC class fractions. Adds overhead to calculation time but useful when \code{subid}
is \code{NULL} or contains many SUBIDs.}
}
\value{
\code{UpstreamPointSources} returns a data frame with columns containing SUBIDs, point source types, volumes, total nitrogen
concentrations, and total phosphorus concentrations.
}
\description{
Function to calculate point source emissions over all upstream areas of a vector of SUBIDs or all SUBIDs in a GeoData table.
Only total nitrogen and total phosphorus emissions are summarised.
}
\details{
\code{UpstreamPointSources} calculates summarised upstream point source emissions. For each sub-basin with at least one upstream
point source (including the sub-basin itself), summed emission volumes and volume weighted emission concentrations are calculated.
HYPE point source types ('ps_type') are returned in separate rows. Water abstraction point sources are ignored
}
\examples{
\dontrun{UpstreamPointSources(subid = 21, gd = mygeodata, psd = mypointsourcedata, bd = mybranchdata)}
}

