% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{WriteAquiferData}
\alias{WriteAquiferData}
\title{Write an 'AquiferData.txt' file}
\usage{
WriteAquiferData(x, filename = "AquiferData.txt", digits = 10,
  nsmall = 0L)
}
\arguments{
\item{x}{The object to be written, a dataframe, as an object returned from \code{\link{ReadAquiferData}}.}

\item{filename}{A character string naming a file to write to. Windows users: Note that
Paths are separated by '/', not '\\'.}

\item{digits}{Integer, number significant digits to export. See \code{\link{format}}.}

\item{nsmall}{Integer, number of significant decimals to export. See \code{\link{format}}.}
}
\description{
This is a convenience wrapper function to export a HYPE 'AquiferData.txt' file from R.
}
\details{
\code{WriteAquiferData} exports a AquiferData dataframe with formatting options adjusted for the output to be read by HYPE.
HYPE requires \code{NA}-free data in all mandatory AquiferData column, but \code{NA}s are allowed in optional comment
columns which are not read by HYPE. The function will return with a warning if \code{NA}s were exported.
}
\examples{
\dontrun{WriteAquiferData(x = myaquiferdata, file = "../myhype/AquiferData.txt")}
}

