% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{WriteBranchData}
\alias{WriteBranchData}
\title{Write a 'BranchData.txt' file}
\usage{
WriteBranchData(x, filename = "BranchData.txt")
}
\arguments{
\item{x}{The object to be written, a dataframe, as an object returned from \code{\link{ReadBranchData}}.}

\item{filename}{A character string naming a file to write to. Windows users: Note that
Paths are separated by '/', not '\\'.}
}
\description{
This is a convenience wrapper function to export a 'BranchData.txt' file from R.
}
\details{
\code{WriteBranchData} exports a BranchData dataframe with formatting options adjusted for the output to be read by HYPE.
HYPE does neither allow empty values in any BranchData column nor any string elements with more than 50 characters, the
function will return with warnings if \code{NA}s or long strings were exported.
}
\examples{
\dontrun{WriteBranchData(x = mybranchdata)}
}

