% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{WriteLakeData}
\alias{WriteLakeData}
\title{Write a 'LakeData.txt' file}
\usage{
WriteLakeData(x, filename = "LakeData.txt", digits = 10, nsmall = 0L)
}
\arguments{
\item{x}{The object to be written, a dataframe, as an object returned from \code{\link{ReadLakeData}}.}

\item{filename}{A character string naming a file to write to. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{digits}{Integer, number significant digits to export. See \code{\link{format}}.}

\item{nsmall}{Integer, number of significant decimals to export. See \code{\link{format}}.}
}
\description{
This is a convenience wrapper function to export a 'LakeData.txt' file from R.
}
\details{
\code{WriteLakeData} exports a LakeData dataframe with formatting options adjusted for the output to be read by HYPE.
HYPE does not allow string elements with more than 50 characters in any LakeData column, the 
function will return with a warning if long strings were exported.
}
\examples{
\dontrun{WriteLakeData(x = mylakedata)}

}

