% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{WriteMapOutput}
\alias{WriteMapOutput}
\title{Write a 'mapXXXX.txt' file}
\usage{
WriteMapOutput(x, filename, dt.format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{The object to be written, a dataframe with \code{comment}, \code{date}, and \code{timestep} 
attributes, as an object returned from \code{\link{ReadMapOutput}}.}

\item{filename}{A character string naming a file to write to. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}. Incomplete format strings for monthly 
and annual values allowed, e.g. '\%Y'.}
}
\description{
Function to export a map output file from R.
}
\details{
\code{WriteTimeOutput} exports a dataframe with headers and formatting options adjusted to match HYPE's map output files.
}
\examples{
\dontrun{WriteMapOutput(x = myCCTN, filename = "mapCCTN.txt")}

}

