% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_export.R
\name{WriteMgmtData}
\alias{WriteMgmtData}
\title{Write a 'MgmtData.txt' file}
\usage{
WriteMgmtData(x, filename = "MgmtData.txt", digits = 10, nsmall = 0L)
}
\arguments{
\item{x}{The object to be written, a dataframe, as an object returned from \code{\link{ReadMgmtData}}.}

\item{filename}{A character string naming a file to write to. Windows users: Note that
Paths are separated by '/', not '\\'.}

\item{digits}{Integer, number significant digits to export. See \code{\link{format}}.}

\item{nsmall}{Integer, number of significant decimals to export. See \code{\link{format}}.}
}
\description{
This is a convenience wrapper function to export a HYPE 'MgmtData.txt' file from R.
}
\details{
\code{WriteMgmtData} exports a MgmtData dataframe with formatting options adjusted for the output to be read by HYPE.
HYPE requires \code{NA}-free data in all mandatory MgmtData column, but \code{NA}s are allowed in optional comment
columns which are not read by HYPE. The function will return with a warning if \code{NA}s were exported.
}
\examples{
\dontrun{WriteMgmtData(x = myMgmtdata, file = "../myhype/MgmtData.txt")}
}

