% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_AllUpstreamSubids.R
\name{AllUpstreamSubids}
\alias{AllUpstreamSubids}
\title{Find All Upstream SUBIDs}
\usage{
AllUpstreamSubids(subid, gd, bd = NULL, write.arcgis = FALSE)
}
\arguments{
\item{subid}{SUBID of a target sub-catchment (must exist in \code{gd}).}

\item{gd}{A data frame, containing 'SUBID' and 'MAINDOWN' columns, e.g. an imported 'GeoData.txt' file. Mandatory argument. See 'Details'.}

\item{bd}{A data frame, containing 'BRANCHID' and 'SOURCEID' columns, e.g. an imported 'BranchData.txt' file. Optional argument.}

\item{write.arcgis}{Logical. If \code{TRUE}, a string containing an SQL expression suitable for ArcGIS's
'Select By Attributes' feature will be written to the clipboard. Works just for Windows.}
}
\value{
\code{AllUpstreamSubids} returns a vector of SUBIDs, ordered as downstream sequence.
}
\description{
Function to find all SUBIDs of upstream sub-catchments for a single
sub-catchment.
}
\details{
\code{AllUpstreamSubids} finds all upstream SUBIDs of a given SUBID (including itself but not
including potential irrigation links or groundwater flows) using GeoData columns 'SUBID' and 'MAINDOWN', i.e the full upstream catchment.
If a BranchData file is provided, the function will also include upstream areas which are connected through an upstream bifurcation. The
results can be directly used as 'partial model setup file' ('pmsf.txt') using the export function \code{\link{WritePmsf}}.
}
\examples{
\dontrun{AllUpstreamSubids(subid = 21, gd = mygeodata)}
}

