% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_AnnualRegime.R
\name{AnnualRegime}
\alias{AnnualRegime}
\title{Calculate annual regimes}
\usage{
AnnualRegime(x, ts.in = NULL, ts.out = NULL, start.mon = 1,
  incl.leap = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame, with column-wise equally-spaced time series. Date-times in \code{\link{POSIXct}} format in first column.
Typically an imported basin or time output file from HYPE. See details.}

\item{ts.in}{Character string, timestep of \code{x}, one of \code{"month"}, \code{"week"}, \code{"day"}, or
\code{"nhour"} (n = number of hours). If not provided, an attribute \code{timestep} is required in \code{x}.}

\item{ts.out}{Character string, timestep for results, defaults to \code{ts.in}. This timestep must be equal to or longer than
\code{ts.in}.}

\item{start.mon}{Integer between 1 and 12, starting month of the hydrological year, used to order the output.}

\item{incl.leap}{Logical, leap days (Mar 29) are removed from results per default, set to \code{TRUE} to keep them. Applies
to daily and shorter time steps only.}

\item{na.rm}{Logical, indicating if \code{NA} values should be stripped before averages are calculated.}
}
\value{
\code{AnnualRegime} returns a list with 5 data frames and a two-element POSIXct vector containing start and end dates of the
source data.
}
\description{
Calculate annual regimes based on long-term time series, typically imported HYPE basin output and time output result files.
}
\details{
\code{AnnualRegime} uses \code{\link{aggregate}} to calculate long-term average regimes for all data columns provided in \code{x},
including long-term arithmetic means, minima and maxima, and 25% and 75% percentiles. In a HYPE context, \code{AnnualRegime}
is particularly applicable to model basin and time results imported using \code{\link{ReadBasinOutput}} and
\code{\link{ReadTimeOutput}}. The function currently does not check if equally spaced time steps are provided in \code{x} or if the
overall time period in \code{x} covers full years so that the calculated averages are based on the same number of values.
}
\note{
If weekly data are provided in \code{x}, \code{AnnualRegime} will inflate \code{x} to daily time steps before computing
results. Values in \code{x} will be assigned to the preceeding week days, corresponding to HYPE file output, where weekly
values are conventionally printed on the last day of the week. If \code{NA} values are present in the original weekly data,
these will be filled with the next available value as a side effect of the inflation.

If weekly output time steps are computed in combination with a user-defined start month, the function will round up weeks when
determining first week of the hydrological year.
}
\examples{
\dontrun{AnnualRegime(x = mybasinoutput)}
}

