% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_BarplotUpstreamClasses.R
\name{BarplotUpstreamClasses}
\alias{BarplotUpstreamClasses}
\title{Bar plots of upstream-averaged classes of HYPE sub-basins}
\usage{
BarplotUpstreamClasses(x, type = NULL, desc = NULL, xlab = NULL,
  ylab = "Area fraction (\%)", ylim = c(0, max(x[, -1] * 1.5)),
  class.names = NULL, names.arg = rep("", ncol(x) - 1), cex.axis = 1,
  cex.names = 0.9, col = NULL, pars = list(mar = c(1.5, 3, 0.5, 0.5) +
  0.1, mgp = c(1.5, 0.3, 0), tcl = NA))
}
\arguments{
\item{x}{Data frame, containing column-wise class group fractions with SUBIDs in first column. Typically a result 
from \code{\link{UpstreamGroupSLCClasses}}.}

\item{type}{Character string keyword for use with \code{desc}. Type of class groups, either \code{"landuse"}, 
\code{"soil"}, or \code{"crop"}, can be abbreviated.}

\item{desc}{List for use with \code{type}. Class description labels imported from a 'description.txt' file, for bar labeling.
See \code{\link{ReadDescription}} for formatting details.}

\item{class.names}{Character vector of class group names, with same length as number of class group fractions in \code{x}. 
Specification of bar labels, alternative to arguments \code{type} and \code{desc}, in particular for cases where a non-standard
grouping was used for \code{x}.}

\item{pars}{}
}
\description{
Function to plot upstream-averaged landscape property classes of one or several sub-basins, e.g. land use or soils, 
as bar plots.
}
\details{
\code{BarplotUpstreamClasses} is a wrapper for \code{\link{barplot}}, with vertical labels plotted over the class group bars.
}
\seealso{
\code{\link{UpstreamGroupSLCClasses}}
\code{\link{barplot}}
}

