% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_HypeAttrAccess.R
\name{HypeAttrAccess}
\alias{HypeAttrAccess}
\alias{datetime}
\alias{obsid}
\alias{outregid}
\alias{subid}
\alias{timestep}
\alias{variable}
\title{Quickly query HYPE-specific attributes}
\usage{
datetime(x)

obsid(x)

outregid(x)

subid(x)

timestep(x)

variable(x)
}
\arguments{
\item{x}{an object whose attribute is to be accessed}
}
\description{
These are simple convenience wrapper functions to quickly query attributes which are added to HYPE files on import.
}
\details{
These functions are just shortcuts for \code{\link{attr}}.
}
