% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{HypeDataImport}
\alias{HypeDataImport}
\alias{ReadAllsim}
\alias{ReadAquiferData}
\alias{ReadBranchData}
\alias{ReadCropData}
\alias{ReadDamData}
\alias{ReadForcKey}
\alias{ReadGlacierData}
\alias{ReadLakeData}
\alias{ReadMgmtData}
\alias{ReadPointSourceData}
\title{Read HYPE data files}
\usage{
ReadAquiferData(filename = "AquiferData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", stringsAsFactors = F, ...)

ReadBranchData(filename = "BranchData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", stringsAsFactors = F, ...)

ReadCropData(filename = "CropData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", stringsAsFactors = F, ...)

ReadDamData(filename = "DamData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", quote = "", stringsAsFactors = F,
  ...)

ReadGlacierData(filename = "GlacierData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", stringsAsFactors = F, ...)

ReadLakeData(filename = "LakeData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", quote = "", stringsAsFactors = F,
  ...)

ReadMgmtData(filename = "MgmtData.txt", verbose = T, header = T,
  na.strings = "-9999", sep = "\\t", stringsAsFactors = F, ...)

ReadPointSourceData(filename = "PointSourceData.txt", verbose = T,
  header = T, na.strings = "-9999", sep = "\\t", stringsAsFactors = F,
  ...)

ReadAllsim(filename = "allsim.txt")

ReadForcKey(filename = "allsim.txt", sep = "\\t")
}
\arguments{
\item{filename}{Path to and file name of HYPE data file file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{verbose}{Logical, display message if columns contain \code{NA} values.}

\item{header}{\code{\link{read.table}} arguments, with appropriote default for HYPE data file import.}

\item{na.strings}{See \code{header}.}

\item{sep}{See \code{header}.}

\item{stringsAsFactors}{See \code{header}.}

\item{...}{Other parameters passed to \code{\link{read.table}}.}

\item{quote}{See \code{header}.}
}
\value{
Imported files are returned as data frames.
}
\description{
These are simple convenience wrapper functions to import various HYPE data files as data frame into R.
}
\details{
Hype data file imports, simple \code{\link{read.table}} wrappers with formatting arguments set to match HYPE file 
specifications:

\itemize{
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:lakedata.txt}{LakeData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:damdata.txt}{DamData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:mgmtdata.txt}{MgmtData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt}{AquiferData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:pointsourcedata.txt}{PointSourceData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:glacierdata.txt}{GlacierData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:cropdata.txt}{CropData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:branchdata.txt}{BranchData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:allsim.txt}{allsim.txt}
}

In most files, HYPE requires \code{NA}-free input in required columns, but empty values are 
allowed in additional comment columns. Informative warnings will be thrown if \code{NA}s are found during import.
}
\examples{
\dontrun{ReadLakeData("LakeData.txt")}
\dontrun{ReadDamData("DamData.txt")}
\dontrun{ReadMgmtData("MgmtData.txt")}

}

