% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{HypeDataImport}
\alias{HypeDataImport}
\alias{ReadAquiferData}
\alias{ReadBranchData}
\alias{ReadCropData}
\alias{ReadDamData}
\alias{ReadGlacierData}
\alias{ReadLakeData}
\alias{ReadMgmtData}
\alias{ReadPointSourceData}
\title{Read HYPE data files}
\usage{
ReadLakeData(filename = "LakeData.txt")

ReadDamData(filename = "DamData.txt")

ReadMgmtData(filename = "MgmtData.txt")

ReadAquiferData(filename = "AquiferData.txt", sep = "\\t")

ReadPointSourceData(filename = "PointSourceData.txt")

ReadGlacierData(filename = "GlacierData.txt", sep = "\\t")

ReadCropData(filename = "CropData.txt")

ReadBranchData(filename = "BranchData.txt")
}
\arguments{
\item{filename}{Path to and file name of HYPE data file file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}
}
\value{
Imported files are returned as data frames.
}
\description{
These are simple convenience wrapper functions to import various HYPE data files as data frame into R.
}
\details{
Hype data file imports, simple \code{\link{read.table}} wrappers with formatting arguments set to match HYPE file 
specifications:

\itemize{
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:lakedata.txt}{LakeData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:damdata.txt}{DamData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:mgmtdata.txt}{MgmtData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt}{AquiferData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:pointsourcedata.txt}{PointSourceData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:glacierdata.txt}{GlacierData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:cropdata.txt}{CropData.txt}
  \item \href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:branchdata.txt}{BranchData.txt}
}

In most files, HYPE requires \code{NA}-free input in required columns, but empty values are 
allowed in additional comment columns. Informative warnings will be thrown if \code{NA}s are found during import.
}
\examples{
\dontrun{ReadLakeData("LakeData.txt")}
\dontrun{ReadDamData("DamData.txt")}
\dontrun{ReadMgmtData("MgmtData.txt")}

}

