% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HypeMultiVar.R
\name{HypeMultiVar}
\alias{HypeMultiVar}
\title{HypeMultiVar arrays}
\usage{
HypeMultiVar(x, date, hype.var, subid = NULL, outregid = NULL)
}
\arguments{
\item{x}{numeric \code{\link{array}} with three dimensions, which holds HYPE results for one sub-basin as (in order)
\code{[date, variable, iteration]}.}

\item{date}{\code{\link{POSIXct}} date-time vector of the same length as \code{time} dimension of \code{x} 
with equidistant time steps (starting day for time steps from weekly to annual), or character string for full model 
period averages, e.g. \code{"2000-2010"}.}

\item{hype.var}{Character vector of keywords to specify HYPE variable IDs, corresponding to second dimension 
(columns) in \code{x}. See 
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:info.txt:variables}{list of HYPE variables}}

\item{subid}{Integer, HYPE sub-basin ID. Either this or \code{outregid} needs to be supplied.}

\item{outregid}{Integer, HYPE output region ID, alternative to \code{subid}.}
}
\value{
Returns a 3-dimensional array with 
\code{[time, variable, iteration]} dimensions and additional \code{\link{attributes}}:
\describe{
\item{\strong{date}}{A vector of date-times. Corresponds to 1st array dimension.}
\item{\strong{variable}}{A character vector of HYPE output variable IDs.}
\item{\strong{subid}}{A single SUBID.}
\item{\strong{subid}}{A single OUTREGID.}
\item{\strong{timestep}}{A character keyword for the time step.}
}
}
\description{
Constructor function for arrays which hold equidistant time series of multiple HYPE variables for a single sub-basin and 
multiple model runs, typically imported HYPE basin output results.
}
\details{
S3 class constructor function for array objects which can hold (multiple) HYPE basin output results.
}
\examples{
\dontrun{HypeMultiVar(mybasinoutput, date = mydates, hype.var = c("cctn", "ccin", "ccon"), , subid = 23, tstep = "day"}

}
