\name{MapRegionalSources}
\alias{MapRegionalSources}
\title{Map regional irrigation source connection as spatial lines}
\usage{
MapRegionalSources(data, map, map.subid.column = 1)
}
\arguments{
  \item{data}{Dataframe with two columns, first column
  containing SUBIDs and second column containing model
  results. See details.}
}
\value{
\code{} returns
}
\description{
This function creates a SpatialLinesDataFrame object which
visualizes regional irrigation source connections between
HYPE subcatchments.
}
\details{
none yet
}
\note{
none yet
}
\examples{
\dontrun{}
}

