% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PlotAnnualRegime.R
\name{PlotAnnualRegime}
\alias{PlotAnnualRegime}
\title{Plot annual regimes}
\usage{
PlotAnnualRegime(x, type = "mean", add.legend = FALSE, l.legend = NULL,
  ylim = NULL, ylab = expression(paste("Q (m"^3, " s"^{     -1 }, ")")),
  xlab = paste(format(x$period, format = "\%Y"), collapse = " to "),
  col = "blue", lty = 1, lwd = 1, mar = c(3, 3, 1, 1) + 0.1)
}
\arguments{
\item{x}{List, typically a result from \code{\link{AnnualRegime}}, containing data frames with aggregated long-term average regime data. 
See details in description there.}

\item{type}{Character string, keyword for plot type. Either \code{"mean"} to plot long-term averages only, or \code{"minmax"} or 
\code{"p25p75"} to include bands of variation. See details.}

\item{add.legend}{Logical. If \code{TRUE}, a legend will be added to the plot.}

\item{l.legend}{Character vector. If non-NULL, legend labels are read from here instead of from column names in \code{x$mean}.}

\item{ylim}{Numeric vector of length two, giving y-axis limits. \code{NULL} for default values.}

\item{ylab}{Character or \code{\link{plotmath}} expression string. Y-axis label, with a default for discharge regimes.}

\item{xlab}{Character string or \code{\link{plotmath}} expression string, x-axis label. Default prints the time period on which the 
regime is based, read from \code{x$period}.}

\item{col}{Line color specification, see \code{\link{par}} for details. Defaults to blue. Either a single value or a vector of the same length as quantile 
series in \code{freq}.}

\item{lty}{Line type specification, see \code{\link{par}} for details. Either a single value or a vector of the same length as quantile 
series in \code{freq}.}

\item{lwd}{Line width specification, see \code{\link{par}} for details. Either a single value or a vector of the same length as quantile 
series in \code{freq}.}

\item{mar}{Numeric vector of length 4, margin specification as in \code{\link{par}} with modified default. Details see there.}
}
\value{
\code{PlotAnnualRegime} returns a plot to the currently active plot device.
}
\description{
Convenience wrapper function for a combined line \code{\link{plot}} with \code{\link{polygon}} variation ranges.
}
\details{
If \code{"minmax"} or \code{"p25p75"} are chosen for argument \code{type}, transparent bands of inter-annual variation are plotted along the 
long-term average line plots, either based on minimum and maximum values or on 25\% and 75\% percentiles.
}
\examples{
\dontrun{PlotAnnualRegime(x = myregime)}
}
\seealso{
\code{\link{AnnualRegime}}
}

