% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_PlotBasinOutput.R
\name{PlotBasinOutput}
\alias{PlotBasinOutput}
\title{Plot a suite of evaluation plots from a HYPE basin output file}
\usage{
PlotBasinOutput(x, timestep = attr(x, "timestep"), log.q = F,
  start.mon = 1, from = 1, to = nrow(x), name = "", area = NULL,
  subid = NULL, gd = NULL, bd = NULL)
}
\arguments{
\item{x}{Data frame, with column-wise equally-spaced time series of HYPE variables. Date-times in
\code{\link{POSIXct}} format in first column. Typically an imported basin output file from HYPE using \code{\link{ReadBasinOutput}}.
See details for HYPE output variables required for plotting.}

\item{timestep}{Character string, timestep of \code{x}, one of \code{"month"}, \code{"week"}, \code{"day"}, or
\code{"nhour"} (n = number of hours). If not provided, an attribute \code{timestep} is required in \code{x}.}

\item{log.q}{}

\item{start.mon}{Integer between 1 and 12, starting month of the hydrological year. For runoff regime plot, see also
\code{\link{AnnualRegime}}.}

\item{from}{}

\item{to}{}

\item{name}{}

\item{area}{Numeric, upstream area of sub-basin in km^2. Required for calculation of accumulated volume error. Optional argument,
either this or arguments \code{subid}, \code{gd}, and \code{bd} are required.}

\item{subid}{HYPE SUBID of a target sub-catchment (must exist in \code{gd}). Mandatory in combination with \code{gd} and
optionally \code{bd} if argument \code{area} is not defined. Used to calculate upstream area internally with function
\code{\link{SumUpstreamArea}}.}

\item{gd}{A data frame, containing 'SUBID' and 'MAINDOWN' columns, e.g. an imported 'GeoData.txt' file. Mandatory with argument
\code{subid}, details see there.}

\item{bd}{A data frame, containing 'BRANCHID' and 'SOURCEID' columns, e.g. an imported 'BranchData.txt' file. Optional with argument
\code{subid}, details see there.}
}
\value{
Returns a multi-panel plot in a new graphics device.
}
\description{
Plot a standard suite of evaluation plots from a basin output file, typically used for model performance inspection and/or
during manual calibration
}
\details{
date, uprf, upsf, temp, uppe, upev, cout, rout, soim, sm13, upfp, snow, uppr, ccin, rein, ccon, reon, cctn, retn, ccsp, resp, ccpp, repp, cctp, retp, wcom, wstr
\url{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:info.txt:variables}
}
\examples{
PlotBasinOutput(x = mybasin, area = 56.67)
}

