% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PlotMapPoints.R
\name{PlotMapPoints}
\alias{PlotMapPoints}
\title{Plot function for mapped point information}
\usage{
PlotMapPoints(x, sites, x.column = 2, sites.subid.column = 1, bg = NULL,
  add = FALSE, map.adj = 0, plot.legend = T, legend.pos = "right",
  legend.title = NULL, legend.inset = c(0, 0), col.ramp.fun = NULL,
  col.breaks = NULL)
}
\description{
Plot mapped point information, e.g. model performances at observation sites.
}

