% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadBasinOutput}
\alias{ReadBasinOutput}
\title{Read a Basin Output File}
\usage{
ReadBasinOutput(filename, dt.format = "\%Y-\%m-\%d", type = "df")
}
\arguments{
\item{filename}{Path to and file name of the basin output file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}. Incomplete format strings for monthly 
and annual values allowed, e.g. '\%Y'. If set to \code{NULL}, no date-time conversion will be attempted and the column will
be imported as \code{character}, applicable e.g. for files containing just one row of summary values over the model period.}

\item{type}{Character, keyword for data type to return. \code{"df"} to return a standard data frame, \code{"dt"} to 
return a \code{\link[data.table]{data.table}} object, or \code{"hmv"} to return a \code{\link{HypeMultiVar}} array.}
}
\value{
\code{ReadBasinOutput} returns a data frame or a matrix, see argument 'outformat'. In the matrix case, date-time information
is converted to numeric POSIX representations (seconds since 1970-01-01). This will lead to NAs if Date-time conversion failed. 
Variable units are imported as string \code{attribute} 'unit' and a time step keyword string in \code{attribute} 'timestep'. 
The catchment's SUBID is extracted from the \code{filename} argument if possible and stored in \code{attribute} 'subid'.
}
\description{
This is a convenience wrapper function to import a basin output file as data frame or matrix into R.
}
\details{
\code{ReadXobs} is a convenience wrapper function of \code{\link[data.table]{fread}} from the 
\code{\link[data.table]{data.table-package}}, with conversion of date-time strings to
POSIX time representations. Monthly and annual time steps are returned as first day of the time step period.
}
\note{
For the conversion of date/time strings, time zone "GMT" is assumed. This is done to avoid potential daylight saving time 
side effects when working with the imported data (and possibly converting to string representations during the process).

HYPE results are printed to files using a user-specified accuracy. This accuracy is specified in 'info.txt' as a number of 
decimals to print. If large numbers are printed, this can result in a total number of digits which is too large to print. 
Results will then contain values of '****************'. \code{ReadBasinOutput} will convert those cases to 'NA' entries.
}
\examples{
\dontrun{ReadBasinOutput("0000001.txt")}

}

