% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadDescription}
\alias{ReadDescription}
\title{Read a 'description.txt' file}
\usage{
ReadDescription(filename, gcl = NULL)
}
\arguments{
\item{filename}{Path to and file name of the 'subassX.txt' file to import.}

\item{gcl}{dataframe, imported GeoClass table to compare number of class names with. A warning will be thrown if numbers 
are not identical.}
}
\value{
\code{ReadDescription} returns a list with 5 elements, corresponding to the imported lines.
}
\description{
Read a 'description.txt' file as \code{list} object into R. A 'description.txt' file contains land use, soil, and crop 
class names of a HYPE set-up, as well as set-up name and version.
}
\details{
\code{ReadDescription} imports a 'description.txt' into R. This file is not used by HYPE, but is convenient for 
e.g. plotting legends or examining imported GeoClass files. The text file must contain 10 lines as follows: HYPE 
set-up name (2nd line), set-up version (4th line), tab-separated land use class names (6th line), tab-separated 
soil class names (8th line),  tab-separated crop class names (8th line). Lines in between can contain headers, 
but are not read by the function.
}
\examples{
\dontrun{ReadDescription("description.txt")}

}

