% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadGeoClass}
\alias{ReadGeoClass}
\title{Read a 'GeoClass.txt' File}
\usage{
ReadGeoClass(filename = "GeoClass.txt")
}
\arguments{
\item{filename}{Path to and file name of the GeoClass file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}
}
\value{
\code{ReadGeoClass} returns a data frame with added attribute 'comment'.
}
\description{
This is a convenience wrapper function to import a GeoClass file as data frame into R. GeoClass files contain definitions
of SLC (\bold{S}oil and \bold{L}and use \bold{C}lass) classes.
}
\details{
\code{ReadGeoClass} is a convenience wrapper function of \code{\link[data.table]{fread}}, with treatment of leading 
comment rows. Column names are created on import, optional comment rows are imported as strings in \code{attribute} 'comment'.
}
\examples{
\dontrun{ReadGeoClass("Geoclass.txt")}

}
