% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadGeoData}
\alias{ReadGeoData}
\title{Read a 'GeoData.txt' file}
\usage{
ReadGeoData(filename = "GeoData.txt", sep = "\\t")
}
\arguments{
\item{filename}{Path to and file name of the GeoData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{sep}{character string. Field separator character as described in \code{\link{read.table}}.}
}
\value{
If the imported file is a HYPE-conform GeoData file, \code{ReadGeoData} returns an object of S3 class \code{\link{HypeGeoData}} 
(see the class description there), providing its own \code{summary} method. If mandatory GeoData columns are missing, 
a standard dataframe is returned along with informative warning messages.
}
\description{
Import a GeoData file into R.
}
\details{
\code{ReadGeoData} uses \code{\link[data.table]{fread}} from the \code{\link{data.table}} package 
with type \code{numeric} type for columns \code{AREA} and \code{RIVLEN} (if they exist), and 
upper-case column names.
}
\examples{
\dontrun{ReadGeoData("GeoData.txt")}

}
