% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadGeoData}
\alias{ReadGeoData}
\title{Read a 'GeoData.txt' file}
\usage{
ReadGeoData(filename = "GeoData.txt", sep = "\\t")
}
\arguments{
\item{filename}{Path to and file name of the GeoData file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{sep}{character string. Field separator character as described in \code{\link{read.table}}.}
}
\value{
\code{ReadGeoData} returns a data frame.
}
\description{
This is a convenience wrapper function to import a GeoData file as data frame into R.
}
\details{
\code{ReadGeoData} runs \code{\link[data.table]{fread}} from the \code{\link{data.table}} 
with type \code{numeric} type for columns \code{AREA} and \code{RIVLEN} (if they exist), and 
upper-case column names.
}
\examples{
\dontrun{ReadGeoData("GeoData.txt")}

}
