% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadMapOutput}
\alias{ReadMapOutput}
\title{Read a Map Output File}
\usage{
ReadMapOutput(filename, colnames = NA)
}
\arguments{
\item{filename}{Path to and file name of the basin output file to import. Windows users: Note that
Paths are separated by '/', not '\\'.}

\item{colnames}{\code{NA} or vector of column names. See Details.}
}
\value{
\code{ReadMapOutput} returns a data frame.
}
\description{
This is a convenience wrapper function to import a map output file ('map<\emph{HYPE_output_variable}>.txt') as data frame into R.
}
\details{
\code{ReadMapOutput} is a convenience wrapper function of \code{\link{read.table}}, with treatment of a leading
comment row in HYPE's output file. The comment row is imported as string \code{attribute} 'comment'. Column names for time period columns
are generated by merging the variable name given in the import file name with column names in the imported file (which are time period
strings). Alternatively, column names can be specified as string vector in the \code{colnames} argument.
}
\note{
HYPE results are printed to files using a user-specified accuracy. This accuracy is specified in 'info.txt' as a number of
decimals to print. If large numbers are printed, this can result in a total number of digits which is too large to print. Results will
then contain values of '****************'. \code{ReadMapOutput} will convert those cases to 'NaN' entries and throw a warning.
}
\examples{
\dontrun{ReadMapOutput("mapCOUT.txt")}
}

