% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadMapOutput}
\alias{ReadMapOutput}
\title{Read a Map Output File}
\usage{
ReadMapOutput(filename, dt.format = NULL)
}
\arguments{
\item{filename}{Path to and file name of the basin output file to import. Windows users: Note that
Paths are separated by '/', not '\\'.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}, for conversion of date-time information in column
headers to POSIX dates, which are returned as attribute. Incomplete format strings for monthly and annual values allowed, e.g.
'\%Y'. Defaults to \code{NULL}, which prevents date-time conversion, applicable e.g. for files containing just one column of
summary values over the model period.}
}
\value{
\code{ReadMapOutput} returns a data frame with additional attributes, see details.
}
\description{
This is a convenience wrapper function to import a map output file ('map<\emph{HYPE_output_variable}>.txt') as data frame into R.
}
\details{
\code{ReadMapOutput} is a convenience wrapper function of \code{\link{read.table}}, with treatment of a leading
comment row in HYPE's output file. The comment row is imported as string \code{attribute} 'comment'. If \code{dt.format} is specified,
POSIX dates are returned in \code{attribute} 'date', and a time step keyword in \code{attribute} 'time step'.
}
\note{
HYPE results are printed to files using a user-specified accuracy. This accuracy is specified in 'info.txt' as a number of
decimals to print. If large numbers are printed, this can result in a total number of digits which is too large to print. Results will
then contain values of '****************'. \code{ReadMapOutput} will convert those cases to 'NaN' entries and throw a warning.
}
\examples{
\dontrun{ReadMapOutput("mapCOUT.txt")}
}

