% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadPTQobs}
\alias{ReadPTQobs}
\title{Read 'Pobs.txt', 'Tobs.txt', 'Qobs.txt', and other observation data files}
\usage{
ReadPTQobs(
  filename,
  variable = c("", "prec", "temp", "rout"),
  dt.format = "\%Y-\%m-\%d",
  nrows = -1,
  type = "df",
  select = NULL,
  obsid = NULL
)
}
\arguments{
\item{filename}{Path to and file name of the file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{variable}{Character string, HYPE variable ID of file contents. If \code{""} (default), the ID is extracted 
from \code{filename}, which only works with file names as required by HYPE or file names beginning with those names 
('e.g. Pobs_old.txt').}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}.}

\item{nrows}{Number of rows to import. A value of \code{-1} indicates all rows, a positive integer gives the number of rows
to import.}

\item{type}{Character, keyword for data type to return. \code{"df"} to return a standard data frame or \code{"dt"} to 
return a \code{\link[data.table]{data.table}} object.}

\item{select}{Integer vector, column numbers to import. Note: first column with dates must be imported and will be added if missing.}

\item{obsid}{Integer vector, HYPE OBSIDs to import. Alternative to argument \code{select}, takes precedence if both are provided.}
}
\value{
\code{ReadPTQobs} returns a data frame with additional attributes: \code{obsid} with observation IDs, \code{timestep} with a time 
step string, either \code{"day"} or \code{"nhour"} (only daily or n-hourly time steps supported), and \code{variable} with a HYPE 
variable ID string.
}
\description{
Import forcing data and discharge observation files as data frame into R.
}
\details{
\code{ReadPTQobs} is a convenience wrapper function of \code{\link[data.table]{fread}} from package  
\code{\link{data.table}}, 
with conversion of date-time strings to POSIX time representations. Observation IDs (SUBIDs or IDs connected to SUBIDs with a 
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:forckey.txt}{ForcKey.txt file}) are returned as integer 
attribute \code{obsid} (see \code{\link{attr}} on how to access it).
}
\note{
For the conversion of date/time strings, time zone "GMT" is assumed. This is done to avoid potential daylight saving time 
side effects when working with the imported data (and e.g. converting to string representations during the process).
}
\examples{
\dontrun{ReadPTQobs("Tobs.txt")}

}
\seealso{
\code{\link{WritePTQobs}}
\code{\link{ReadXobs}}
}
