% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadTimeOutput}
\alias{ReadTimeOutput}
\title{Read a Time Output File}
\usage{
ReadTimeOutput(
  filename,
  dt.format = "\%Y-\%m-\%d",
  hype.var = NULL,
  type = "df",
  select = NULL,
  subid = NULL,
  nrows = -1L,
  skip = 0L,
  warn.nan = FALSE
)
}
\arguments{
\item{filename}{Path to and file name of the time output file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}. Incomplete format strings for monthly 
and annual values allowed, e.g. '\%Y'. If set to \code{NULL}, no date-time conversion will be attempted and the column will
be imported as \code{character}, applicable e.g. for files containing just one row of summary values over the model period.}

\item{hype.var}{Character vector of four-letter keywords to specify HYPE variable IDs, corresponding to second dimension 
(columns) in \code{x}. See 
\href{http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:info.txt:variables}{list of HYPE variables}.
If \code{NULL} (default), the variable ID is extracted from the provided file name, which only works for standard HYPE 
time output file names.}

\item{type}{Character, keyword for data type to return. \code{"df"} to return a standard data frame, \code{"dt"} to 
return a \code{\link[data.table]{data.table}} object, or \code{"hsv"} to return a \code{\link{HypeSingleVar}} array.}

\item{select}{Integer vector, column numbers to import. Note: first column with dates must be imported and will be added if missing.}

\item{subid}{Integer vector, HYPE SUBIDs to import. Alternative to argument \code{select}, takes precedence if both are provided.}

\item{nrows}{Integer, number of rows to import, see documentation in \code{\link[data.table]{fread}}.}

\item{skip}{Integer, number of \strong{data} rows to skip on import. Time output header lines are always skipped.}

\item{warn.nan}{Logical, check if imported results contain any \code{NaN} values. If \code{TRUE} and \code{NaN}s are found, 
a warning is thrown and affected SUBIDs saved in an attribute \code{subid.nan}. Adds noticeable overhead to import time for large files.}
}
\value{
\code{ReadTimeOutput} returns a \code{data.frame}, \code{\link{data.table}}, or a \code{\link{HypeSingleVar}} array. 
Data frames and data tables contain additional \code{\link{attributes}}: \code{variable}, giving the HYPE variable ID, 
\code{subid}, a vector of subid integers (corresponding to columns from column 2), \code{timestep} with a time step attribute, 
and \code{comment} with first row comment of imported file as character string. An additional attribute \code{subid.nan} might be 
returned, see argument \code{warn.nan}.
}
\description{
This is a convenience wrapper function to import a time output file ('time<\emph{HYPE_output_variable}>.txt') into R.
}
\details{
\code{ReadTimeOutput} is a convenience wrapper function of \code{\link[data.table]{fread}} from package  
\code{\link{data.table}}, 
with conversion of date-time strings to POSIX time representations. Monthly and annual time steps are returned as first day 
of the time step period.
}
\note{
For the conversion of date/time strings, time zone "GMT" is assumed. This is done to avoid potential daylight saving time 
side effects when working with the imported data (and possibly converting to string representations during the process).

HYPE results are printed to files using a user-specified accuracy. This accuracy is specified in 'info.txt' as a number of 
decimals to print. If large numbers are printed, this can result in a total number of digits which is too large to print. 
Results will then contain values of '****************'. \code{ReadTimeOutput} will convert those cases to 'NA' entries.

Current versions of HYPE allow for defining significant instead of fixed number of digits, which should prevent this 
issue from arising.
}
\examples{
\dontrun{ReadTimeOutput("timeCCIN.txt", dt.format = "\%Y-\%m")}

}
