% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncollection_import.R
\name{ReadXobs}
\alias{ReadXobs}
\title{Read an 'Xobs.txt' file}
\usage{
ReadXobs(filename = "Xobs.txt", dt.format = "\%Y-\%m-\%d", nrows = -1)
}
\arguments{
\item{filename}{Path to and file name of the Xobs file to import. Windows users: Note that 
Paths are separated by '/', not '\\'.}

\item{dt.format}{Date-time \code{format} string as in \code{\link{strptime}}.}

\item{nrows}{Integer, number of rows to import. A value of \code{-1} indicates all rows, a positive integer gives the number of rows
to import.}
}
\value{
If datetime import to POSIXct worked, \code{ReadXobs} returns a \code{\link{HydroXobs}} object, a data frame with three 
additional attributes \code{variable}, \code{subid}, and \code{comment}: \code{variable} 
and \code{subid} each contain a vector with column-wise HYPE IDs (first column with date/time information omitted). 
\code{comment} contains the content of the Xobs file comment row as single string. Column names of the returned data 
frame are composed of variable names and SUBIDs, separated by and underscore, i.e. \code{[variable]_[subid]}. 
If datetime conversion failed on import, the returned object is a data frame (i.e. no class \code{HypeXobs}).
}
\description{
This is a convenience wrapper function to import an Xobs file into R.
}
\details{
\code{ReadXobs} is a convenience wrapper function of \code{\link{read.table}}, with conversion of date-time strings to
POSIX time representations. Variable names, SUBIDs, and comments are returned as attributes (see \code{\link{attr}} on 
how to access these).
}
\note{
For the conversion of date/time strings, time zone "GMT" is assumed. This is done to avoid potential daylight saving time 
side effects when working with the imported data (and possibly converting to string representations during the process).
}
\examples{
\dontrun{ReadXobs("Xobs.txt")}

}

