% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/function_UpstreamGeoData.R
\name{UpstreamGeoData}
\alias{UpstreamGeoData}
\title{Calculate upstream sums and averages of selected GeoData contents}
\usage{
UpstreamGeoData(subid = NULL, gd, bd = NULL, olake.slc = NULL,
  signif.digits = 3, progbar = TRUE)
}
\arguments{
\item{subid}{Integer vector of SUBIDs for which to calculate upstream properties (must exist in \code{gd}).
If \code{NULL} (default), upstream areas for all SUBIDs will be calculated.}

\item{gd}{A data frame containing a column with SUBIDs and a column with areas, e.g. an imported 'GeoData.txt' file.}

\item{bd}{A data frame with bifurcation connections, e.g. an imported 'BranchData.txt' file. Optional argument.}

\item{olake.slc}{Integer,SLC class number which represents outlet lake fractions. Mandatory for weighted averaging of outlet lake depths.}

\item{signif.digits}{Integer, number of significant digits to round upstream SLCs to. See also \code{\link{signif}}. Set to \code{NULL} to prevent rounding.}

\item{progbar}{Logical, display a progress bar while calculating SLC class fractions. Adds overhead to calculation time but useful when \code{subid}
is \code{NULL} or contains many SUBIDs.}
}
\value{
\code{UpstreamGeoData} returns a data frame with the same number of columns as argument \code{gd} and number of rows corresponding to number of
SUBIDs in argument \code{subid}, with updated upstream columns marked with a leading 'UP_' in the column names.
}
\description{
Function to calculate upstream sums and averages for selected variables of imported GeoData.txt files.
}
\details{
\code{UpstreamGeoData} calculates upstream averages or sums of selected variables in a GeoData data frame, including branch connections
in case of stream bifurcations but not including potential irrigation links or groundwater flows. Averages are weighted by sub-catchment area, with
the exception of outlet lake depths provided in GeoData variable 'lake_depth'. These are weighted by outlet lake area and the GeoData column with
SLC class fractions for outlet lakes must be provided in function argument \code{col.olake.slc}. Elevation and slope standard deviations are
averaged if the corresponding mean values exist (sample means are required to calculate overall means of std. devs.).

Currently, the following variables are considered:
\describe{
  \item{Area-weighted average}{elev_mean, slope_mean, buffer, close_w, latitude, longitude, all SLC classes, elev_std, slope_std}
  \item{Sum}{area, rivlen}
}
}
\examples{
\dontrun{UpstreamGeoData(subid = 21, gd = mygeodata, bd = mybranchdata, olake.slc = 1)}
}
\seealso{
\code{\link{UpstreamSLCClasses}}
\code{\link{SumUpstreamArea}}
\code{\link{AllUpstreamSubids}}
}

